move pub use to the top, keep only one way of reaching reexports
